+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 0.3.1
+
+* Add globOptions to generator file helpers
+* Add globOptions to BattleCry class (used on battlecry-setup.js)
+* Set default globOptions to allow listing files starting with `.` except for `._DS_Store`
+
 # 0.3
 
 * Variadic optional arguments (`...arg?`) become null instead of an empty array when no values are given